requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
mysql-connector-python==9.0.0
scikit-learn==1.5.1
matplotlib==3.8.0
plotly==5.23.0
bcrypt==4.2.0
google-generativeai==0.7.2
streamlit-authenticator==0.3.3
fastapi==0.68.2
uvicorn==0.30.6